+++ next.config.js
@@ -54,7 +54,7 @@ const nextConfig = {
   reactStrictMode: true,
   trailingSlash: true,
   env: {
-    version: '3.7.10'
+    version: '3.7.12'
   },
   devIndicators: {
     buildActivity: false